Create Elec Desc Maker.py
This script was created to automate the naming process of hundreds of assets soon to be within our system.

Certain letters in the asset ID stand for different phrases that needed to be spelled out in the description box. This script dissects the asset IDs and writes out the description depending on what values each section holds. This script also assigns 2 other roles to each asset, the class and position.